refactor: XIcon -> CloseIcon 파일명 변경
